Split `groupsManager` and `group`
- only blocker was unused `get_all_groups`
- includes more documentation
